Adressed some review comments
